+++ psyneulink/core/components/functions/__init__.py
@@ -6,6 +6,7 @@ from . import transferfunctions
 from . import selectionfunctions
 from . import integratorfunctions
 from . import distributionfunctions
+from . import objectivefunctions
 
 from .function import *
 from .userdefinedfunction import *
@@ -15,6 +16,7 @@ from .transferfunctions import *
 from .selectionfunctions import *
 from .integratorfunctions import *
 from .distributionfunctions import *
+from .objectivefunctions import *
 
 __all__ = list(function.__all__)
 __all__.extend(userdefinedfunction.__all__)
@@ -24,3 +26,4 @@ __all__.extend(transferfunctions.__all__)
 __all__.extend(selectionfunctions.__all__)
 __all__.extend(integratorfunctions.__all__)
 __all__.extend(distributionfunctions.__all__)
+__all__.extend(objectivefunctions.__all__)
